Improve Rat King loots.

+++ itsyscape/Resources/Game/Maps/HighChambersYendor_Floor2/Peep.lua
@@ -11,6 +11,8 @@ local Class = require "ItsyScape.Common.Class"
 local Utility = require "ItsyScape.Game.Utility"
 local Map = require "ItsyScape.Peep.Peeps.Map"
 local Probe = require "ItsyScape.Peep.Probe"
+local CombatStatusBehavior = require "ItsyScape.Peep.Behaviors.CombatStatusBehavior"
+local PlayerBehavior = require "ItsyScape.Peep.Behaviors.PlayerBehavior"
 local HighChambersYendorCommon = require "Resources.Game.Peeps.HighChambersYendor.Common"
 
 local HighChambersYendor = Class(Map)
@@ -31,13 +33,13 @@ function HighChambersYendor:initMiniboss()
 		self:getLayerName(),
 		Probe.resource("Peep", "HighChambersYendor_RatKing"))
 	if #hits >= 1 then
-		hits[1]:listen('die', self.giveMinibossLoot, self)
+		hits[1]:listen('die', self.giveMinibossLoot, self, hits[1])
 
 		Log.info("Found Rat King miniboss.")
 	end
 end
 
-function HighChambersYendor:giveMinibossLoot()
+function HighChambersYendor:giveMinibossLoot(ratKing)
 	local director = self:getDirector()
 	local gameDB = director:getGameDB()
 
@@ -52,14 +54,20 @@ function HighChambersYendor:giveMinibossLoot()
 			Probe.mapObject(chest:get("Resource")))[1]
 
 		if chest then
-			chest:poke('materialize', {
-				count = math.random(20, 40),
-				dropTable = gameDB:getResource("HighChambersYendor_RatKing_Rewards", "DropTable"),
-				peep = director:getGameInstance():getPlayer():getActor():getPeep(),
-				chest = chest
-			})
+			local status = ratKing:getBehavior(CombatStatusBehavior)
+			for peep in pairs(status.damage) do
+				local isPlayer = peep:hasBehavior(PlayerBehavior)
+				if isPlayer then
+					chest:poke('materialize', {
+						count = math.random(20, 40),
+						dropTable = gameDB:getResource("HighChambersYendor_RatKing_Rewards", "DropTable"),
+						peep = peep,
+						chest = chest
+					})
 
-			Log.info("Rat King loot materialized.")
+					Log.info("Rat King loot materialized for player '%s'.", peep:getName())
+				end
+			end
 		end
 	end
 end
